Extracted classes KeyGrabber and Recorder from iomediator00 into a separate module.
