+++ lib/autokey/iomediator/__init__.py
@@ -19,3 +19,4 @@ IoMediator public interface
 
 from .windowgrabber import WindowGrabber
 from ._waiter import Waiter
+from ._keygrabber import Recorder, KeyGrabber
