+++ lib/autokey/qtui/dialogs.py
@@ -23,11 +23,11 @@ from PyKDE4.kdecore import i18n
 from PyQt4.QtGui import *
 from PyQt4.QtCore import Qt
 
+
 __all__ = ["validate", "EMPTY_FIELD_REGEX", "AbbrSettingsDialog", "HotkeySettingsDialog", "WindowFilterSettingsDialog", "RecordDialog"]
 
 from . import abbrsettings, hotkeysettings, windowfiltersettings, recorddialog, detectdialog
 from .. import model, iomediator
-from ..iomediator import iomediator00
 from ..iomediator.key import Key
 
 WORD_CHAR_OPTIONS = {
@@ -59,12 +59,12 @@ class AbbrListItem(QListWidgetItem):
             QListWidgetItem.setData(self, role, value)
 
 class AbbrSettings(QWidget, abbrsettings.Ui_Form):
-    
+
     def __init__(self, parent):
         QWidget.__init__(self, parent)
         abbrsettings.Ui_Form.__init__(self)
         self.setupUi(self)
-        
+
         for item in WORD_CHAR_OPTIONS_ORDERED:
             self.wordCharCombo.addItem(item)
 
@@ -87,21 +87,21 @@ class AbbrSettings(QWidget, abbrsettings.Ui_Form):
             row = self.abbrListWidget.row(item)
             self.abbrListWidget.takeItem(row)
             del item
-            
+
         if self.abbrListWidget.count() == 0:
             self.removeButton.setEnabled(False)
 
     def on_abbrListWidget_itemDoubleClicked(self, item):
         self.abbrListWidget.editItem(item)
-        
+
     def on_ignoreCaseCheckbox_stateChanged(self, state):
         if not self.ignoreCaseCheckbox.isChecked():
             self.matchCaseCheckbox.setChecked(False)
-            
+
     def on_matchCaseCheckbox_stateChanged(self, state):
         if self.matchCaseCheckbox.isChecked():
             self.ignoreCaseCheckbox.setChecked(True)
-            
+
     def on_immediateCheckbox_stateChanged(self, state):
         if self.immediateCheckbox.isChecked():
             self.omitTriggerCheckbox.setChecked(False)
@@ -122,7 +122,7 @@ class AbbrSettingsDialog(KDialog):
         self.setPlainCaption(i18n("Set Abbreviations"))
         self.setModal(True)
         #self.connect(self, SIGNAL("okClicked()"), self.on_okClicked)
-        
+
     def load(self, item):
         self.targetItem = item
         self.widget.abbrListWidget.clear()
@@ -134,7 +134,7 @@ class AbbrSettingsDialog(KDialog):
             self.widget.abbrListWidget.setCurrentRow(0)
         else:
             self.widget.removeButton.setEnabled(False)
-        
+
         self.widget.removeTypedCheckbox.setChecked(item.backspace)
 
         self.__resetWordCharCombo()
@@ -150,46 +150,46 @@ class AbbrSettingsDialog(KDialog):
             # Custom wordchar regex used
             self.widget.wordCharCombo.addItem(model.extract_wordchars(wordCharRegex))
             self.widget.wordCharCombo.setCurrentIndex(len(WORD_CHAR_OPTIONS))
-        
+
         if isinstance(item, model.Folder):
             self.widget.omitTriggerCheckbox.setVisible(False)
         else:
             self.widget.omitTriggerCheckbox.setVisible(True)
             self.widget.omitTriggerCheckbox.setChecked(item.omitTrigger)
-        
+
         if isinstance(item, model.Phrase):
             self.widget.matchCaseCheckbox.setVisible(True)
             self.widget.matchCaseCheckbox.setChecked(item.matchCase)
         else:
             self.widget.matchCaseCheckbox.setVisible(False)
-        
+
         self.widget.ignoreCaseCheckbox.setChecked(item.ignoreCase)
         self.widget.triggerInsideCheckbox.setChecked(item.triggerInside)
         self.widget.immediateCheckbox.setChecked(item.immediate)
-        
+
     def save(self, item):
         item.modes.append(model.TriggerMode.ABBREVIATION)
         item.clear_abbreviations()
         item.abbreviations = self.get_abbrs()
-        
+
         item.backspace = self.widget.removeTypedCheckbox.isChecked()
-        
+
         option = str(self.widget.wordCharCombo.currentText())
         if option in WORD_CHAR_OPTIONS:
             item.set_word_chars(WORD_CHAR_OPTIONS[option])
         else:
             item.set_word_chars(model.make_wordchar_re(option))
-        
+
         if not isinstance(item, model.Folder):
             item.omitTrigger = self.widget.omitTriggerCheckbox.isChecked()
-            
+
         if isinstance(item, model.Phrase):
             item.matchCase = self.widget.matchCaseCheckbox.isChecked()
-            
+
         item.ignoreCase = self.widget.ignoreCaseCheckbox.isChecked()
         item.triggerInside = self.widget.triggerInsideCheckbox.isChecked()
         item.immediate = self.widget.immediateCheckbox.isChecked()
-        
+
     def reset(self):
         self.widget.removeButton.setEnabled(False)
         self.widget.abbrListWidget.clear()
@@ -206,13 +206,13 @@ class AbbrSettingsDialog(KDialog):
         for item in WORD_CHAR_OPTIONS_ORDERED:
             self.widget.wordCharCombo.addItem(item)
         self.widget.wordCharCombo.setCurrentIndex(0)
-        
+
     def get_abbrs(self):
         ret = []
         for i in range(self.widget.abbrListWidget.count()):
             text = self.widget.abbrListWidget.item(i).text()
             ret.append(str(text))
-            
+
         return ret
 
     def get_abbrs_readable(self):
@@ -230,7 +230,7 @@ class AbbrSettingsDialog(KDialog):
                             self.widget.addButton, self): return False
 
         return True
-        
+
     def slotButtonClicked(self, button):
         if button == KDialog.Ok:
             if self.__valid():
@@ -241,18 +241,18 @@ class AbbrSettingsDialog(KDialog):
 
 
 class HotkeySettings(QWidget, hotkeysettings.Ui_Form):
-    
+
     def __init__(self, parent):
         QWidget.__init__(self, parent)
         hotkeysettings.Ui_Form.__init__(self)
-        self.setupUi(self)    
+        self.setupUi(self)
 
     # ---- Signal handlers
-    
+
     def on_setButton_pressed(self):
         self.setButton.setEnabled(False)
         self.keyLabel.setText(i18n("Press a key or combination..."))
-        self.grabber = iomediator00.KeyGrabber(self.parentWidget())
+        self.grabber = iomediator.KeyGrabber(self.parentWidget())
         self.grabber.start()  
 
 class HotkeySettingsDialog(KDialog):
